+++ src/components/Utils/orderResource.js
@@ -6,8 +6,6 @@ const saveOrder = (order, mutator) => {
   delete order.createdByName;
   delete order.assignedToUser;
   delete order.vendorName;
-  delete order.numberPrefix;
-  delete order.numberSuffix;
 
   if (order.id) {
     method = mutator.PUT;
@@ -17,8 +15,14 @@ const saveOrder = (order, mutator) => {
   return method(order);
 };
 
+export const getFullOrderNumber = ({ poNumberPrefix = '', poNumberSuffix = '', poNumber = '' } = {}) => (
+  `${poNumberPrefix}${poNumber}${poNumberSuffix}`.trim()
+);
+
 export const updateOrderResource = (order, mutator, changedProps) => {
   const clonedOrder = cloneDeep(order);
+
+  clonedOrder.poNumber = getFullOrderNumber(clonedOrder);
 
   Object.assign(clonedOrder, changedProps);
 
@@ -27,16 +31,17 @@ export const updateOrderResource = (order, mutator, changedProps) => {
 
 export const createOrderResource = (order, mutator) => {
   const clonedOrder = cloneDeep(order);
-  const { numberPrefix = '', numberSuffix = '', poNumber = '' } = clonedOrder;
-  const fullOrderNumber = `${numberPrefix}${poNumber}${numberSuffix}`.trim();
 
-  clonedOrder.poNumber = fullOrderNumber || undefined;
+  clonedOrder.poNumber = getFullOrderNumber(clonedOrder) || undefined;
 
   return saveOrder(clonedOrder, mutator);
 };
 
 export const cloneOrder = (order, mutator, lines) => {
-  const clonedOrder = omit(order, ['id', 'adjustment', 'metadata', 'poNumber', 'workflowStatus', 'compositePoLines']);
+  const clonedOrder = omit(
+    order,
+    ['id', 'adjustment', 'metadata', 'poNumber', 'poNumberPrefix', 'poNumberSuffix', 'workflowStatus', 'compositePoLines'],
+  );
 
   if (lines) {
     clonedOrder.compositePoLines = lines.map(line => omit(line, ['id', 'purchaseOrderId', 'metadata']));